competitions/kaggle_inteisve_ai/pyproject.toml
[project]
name = "emergencycarenavigator"
version = "1.0.0"
description = "EmergencyCareNavigator API"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "requests==2.31.0",
    "pydantic==2.5.0",
    "google-generativeai==0.3.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "PyPDF2==3.0.1",
    "pytesseract==0.3.10",
    "Pillow==10.1.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# Tell setuptools explicitly which Python packages to include so it doesn't
# try to auto-discover every top-level folder (frontend, static, uploads, etc.)
[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["frontend*", "static*", "uploads*", "docs*", "venv*", "*.ipynb", "api*"]

[tool.vercel]
app = "api.app:app"



